+++ src/main/java/org/multicoder/mcpaintball/event/MCPaintballGameEvents.java
@@ -1,28 +1,25 @@
 package org.multicoder.mcpaintball.event;
 
-import com.mojang.brigadier.CommandDispatcher;
+
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screens.Screen;
-import net.minecraft.commands.CommandBuildContext;
-import net.minecraft.commands.CommandSourceStack;
-import net.minecraft.commands.Commands;
-import net.minecraft.network.chat.Component;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.server.permissions.Permissions;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.entity.player.Player;
+import org.multicoder.mcpaintball.client.gui.AdminScreen;
 import org.multicoder.mcpaintball.client.gui.SelectRoleScreen;
 import org.multicoder.mcpaintball.client.gui.SelectTeamScreen;
-import org.multicoder.mcpaintball.command.MCPaintballCommands;
 import org.multicoder.mcpaintball.core.MCPaintballDataAttachments;
 import org.multicoder.mcpaintball.core.MCPaintballDataComponents;
 import org.multicoder.mcpaintball.core.MCPaintballItems;
 import org.multicoder.mcpaintball.core.MCPaintballKeybinding;
 import org.multicoder.mcpaintball.data.MCPaintballSaveData;
 import org.multicoder.mcpaintball.network.CycleGLTypeC2SPacket;
+import org.multicoder.mcpaintball.network.KitSenderC2SPacket;
 import org.multicoder.mcpaintball.network.PointSyncS2CPacket;
 import org.multicoder.mcpaintball.integration.MinecraftTeamSystem;
 
@@ -49,12 +46,6 @@ public class MCPaintballGameEvents {
         player.getAttachedOrCreate(MCPaintballDataAttachments.PAINTBALL_PLAYER);
     }
 
-    public static void commandRegister(CommandDispatcher<CommandSourceStack> dispatcher, CommandBuildContext ignored, Commands.CommandSelection ignored2) {
-        dispatcher.register(Commands.literal(Component.translatable("command.mcpaintball.prefix").getString()).then(Commands.literal(Component.translatable("command.mcpaintball.game_prefix").getString()).then(Commands.literal(Component.translatable("command.mcpaintball.start").getString()).requires(commandSourceStack -> commandSourceStack.permissions().hasPermission(Permissions.COMMANDS_ADMIN)).executes(MCPaintballCommands::startGame)).then(Commands.literal(Component.translatable("command.mcpaintball.stop").getString()).requires(commandSourceStack -> commandSourceStack.permissions().hasPermission(Permissions.COMMANDS_ADMIN)).executes(MCPaintballCommands::stopGame)))).createBuilder().build();
-        dispatcher.register(Commands.literal(Component.translatable("command.mcpaintball.prefix").getString()).then(Commands.literal(Component.translatable("command.mcpaintball.round_prefix").getString()).then(Commands.literal(Component.translatable("command.mcpaintball.start").getString()).requires(commandSourceStack -> commandSourceStack.permissions().hasPermission(Permissions.COMMANDS_ADMIN)).executes(MCPaintballCommands::startRound)).then(Commands.literal(Component.translatable("command.mcpaintball.end").getString()).requires(commandSourceStack -> commandSourceStack.permissions().hasPermission(Permissions.COMMANDS_ADMIN)).executes(MCPaintballCommands::stopRound)).then(Commands.literal(Component.translatable("command.mcpaintball.winner").getString()).requires(commandSourceStack -> commandSourceStack.permissions().hasPermission(Permissions.COMMANDS_MODERATOR)).executes(MCPaintballCommands::roundWinner)))).createBuilder().build();
-        dispatcher.register(Commands.literal(Component.translatable("command.mcpaintball.prefix").getString()).then(Commands.literal(Component.translatable("command.mcpaintball.kit").getString()).executes(MCPaintballCommands::giveKit))).createBuilder().build();
-    }
-
     public static void clientEndTick(Minecraft minecraft) {
         while(MCPaintballKeybinding.CYCLE_GRENADE_LAUNCHER_TYPE.consumeClick()){
             Player player =  minecraft.player;
@@ -70,6 +61,15 @@ public class MCPaintballGameEvents {
         while (MCPaintballKeybinding.OPEN_ROLE_SELECT.consumeClick()){
             Screen Parent = minecraft.screen;
             minecraft.setScreen(new SelectRoleScreen(Parent));
+        }
+        while (MCPaintballKeybinding.OPEN_ADMIN_SCREEN.consumeClick()){
+            if(Objects.requireNonNull(minecraft.player).permissions().hasPermission(Permissions.COMMANDS_MODERATOR)){
+                Screen Parent = minecraft.screen;
+                minecraft.setScreen(new AdminScreen(Parent));
+            }
+        }
+        while (MCPaintballKeybinding.KIT_GIVE.consumeClick()){
+            ClientPlayNetworking.send(new KitSenderC2SPacket());
         }
     }
 }